manager as proxy for controllers
